Move all Morph stuff into the Profile class. Add more docstrings
